feat(https): working https!
Some things had to change. The most important change, I believe, was
changing the external url from gitlab.abarnett.com to the ip of the
docker machine. Since I couldn't do any dns configuration,
gitlab.abarnett.com never led to anything. I thought I needed the port
in the external url, but it didn't work when I had that in. I've also
changed the location of the ssl certificates, but I'm not sure if that
was one of the changes that mattered. I saw some documentation that said
they should go there, instead of where I had them before (which, funny
enough, I also found in documentation). I've also changed it to do all
of the reconfiguring when the container is launched, not when it builds.
I might change it back, I liked how it was before. I added configuration
for the ssh port, but ssh still isn't working. I am currently working on
setting up the ssh daemon in the gitlab container with runit. I've also
added some more resources to the dockerfile.