feat(ssh!!!): ssh works!
Happy day happy day! Everything's done finally. This commit adds in a
snippet that does some pseudo-servicing to actually make sure the ssh
daemon is running and listening, because it was not before. I also read
over the dockerfiles best practices guide, and cleaned up my apt-getting
at the top of the file.